refactor the utility function to convert manual positions
