+++ BatteryLab/helper/utils.py
@@ -4,6 +4,9 @@ import pyudev # Linux only
 from pathlib import Path
 from serial.tools import list_ports
 from enum import Enum
+import numpy as np
+from typing import List
+from ..robots.Constants import AssemblyRobotConstants, ComponentProperty
 
 class SupportedDevices(Enum):
     ZaberLinearRail = 1
@@ -55,3 +58,77 @@ def get_proper_port_for_device(device_name: SupportedDevices):
         selected_port = usb_ports[port_index].device
     print("selected port: ", selected_port)
     return selected_port
+
+
+def get_8_8_well_pos(bottom_left_coordinates, bottom_right_coordinates, top_left_coordinates, top_right_coordinates):
+    avg_height = np.average([bottom_left_coordinates[2], bottom_right_coordinates[2], top_left_coordinates[2], top_right_coordinates[2]])
+    pos_dict: dict[int, List[float]] = {}
+    delta_y = np.average([bottom_right_coordinates[1] - bottom_left_coordinates[1], top_right_coordinates[1] - top_left_coordinates[1]]) / 8
+    delta_x = np.average([bottom_right_coordinates[0] - top_right_coordinates[0], bottom_left_coordinates[0] - top_left_coordinates[0]]) / 8
+    for i in range(64):
+        x = i / 8
+        y = i % 8
+        pos_dict[i] = [top_left_coordinates[0] + x * delta_x, top_left_coordinates[1] + y * delta_y, avg_height] + bottom_left_coordinates[3:]
+    return pos_dict
+
+def create_robot_constants_from_manual_positions(manual_positions) -> AssemblyRobotConstants:
+    """
+    Obtain the assembly robot constants from manually aligned positions.
+
+    Parameters:
+      manual_positions: the YAML file content parsed by yaml.safe_load()
+
+    Returns:
+      AssemblyRobotConstants: generate all the well positions and 
+
+    There are 7 fields for different components, namely Cathode Case, Cathode, Separator, Anode, Washer, Spacer, Anode Case.
+    Each component tray will have 4 manually posed positions at 4 corners (x, y, z, alpha, beta, gama). There are 64 wells on each tray.
+    The location for each well will be bilinearly interpolated by the 4 positions.
+
+    An example YAML file looks like the following.
+
+AessemblePost:
+  rail_pos: 110
+  cartesian: [-193.99646, -152.00122, 126.08627, 163.49303, -50.50488, -21.00815]
+  joints: [-135.12872, 42.12607, 7.25795, -105.43295, 25.59059, 127.64982]
+Separator:
+  rail_pos: 14.5
+  bottom_left:
+    cartesian: [235.97647, -87.38851, 40.11804, 173.25027, 42.36252, -166.87105]
+    joints: [-23.29881, 67.05935, 0.30938, -30.22345, -21.77718, 19.73194]
+  bottom_right:
+    cartesian: [235.16858, 72.84175, 39.81935, 173.25028, 42.36252, -166.87105]
+    joints: [22.77213, 66.685, 1.47282, 51.36269, -25.87523, -22.1584]
+  top_left:
+    cartesian: [75.24227, -87.65825, 41.29566, 173.25028, 42.36252, -166.87105]
+    joints: [-71.00944, 52.73619, 64.5154, -45.23412, -59.04767, -25.42198]
+  top_right:
+    cartesian: [74.08437, 72.84175, 41.03251, 173.25028, 42.36252, -166.87105]
+    joints: [71.16736, 53.23022, 65.68386, 54.59531, -54.76851, 38.62172]
+Cathode:
+    ...
+    """
+    constants = AssemblyRobotConstants()
+
+    cartesian_coord_prop = "cartesians"
+    rail_pos_prop = "rail_pos"
+    bottom_left_prop = "bottom_left"
+    bottom_right_prop = "bottom_right"
+    top_left_prop = "top_left"
+    top_right_prop = "top_right"
+
+    assemble_post = manual_positions["AssemblePost"]
+    constants.POST_C_SK_PO = assemble_post[cartesian_coord_prop]
+    constants.POST_RAIL_LOCATION = assemble_post[rail_pos_prop]
+
+    separator = manual_positions["Separator"]
+    constants.Separator = ComponentProperty()
+    constants.Separator.railPo = separator[rail_pos_prop]
+    constants.Separator.dropPo = assemble_post[cartesian_coord_prop]
+    bottom_left_coordinates = separator[bottom_left_prop][cartesian_coord_prop]
+    bottom_right_coordinates = separator[bottom_right_prop][cartesian_coord_prop]
+    top_left_coordinates = separator[top_left_prop][cartesian_coord_prop]
+    top_right_coordinates = separator[top_right_prop][cartesian_coord_prop]
+    constants.Separator.grabPo = get_8_8_well_pos(bottom_left_coordinates, bottom_right_coordinates, top_left_coordinates, top_right_coordinates)
+
+    return constants
