+++ ros2_ws/src/assembly_robot/assembly_robot/test_yaml_path.py
@@ -1,9 +1,11 @@
 import yaml
+import rclpy
 from pathlib import Path
+from ament_index_python.packages import get_package_share_path
 
 def main():
-    
-    position_file = Path(__file__).parent / ".." / "resource" / "well_positions.yaml"
+    rclpy.init()
+    position_file = Path(get_package_share_path("assembly_robot")) / "yaml" / "well_positions.yaml"
 
     with open(position_file, "r") as f:
         try:
@@ -15,6 +17,7 @@ def main():
     assemble_post = constant_positions["AessemblePost"]
     print(assemble_post["rail_pos"])
     print(assemble_post["cartesian"])
+    rclpy.shutdown()
 
 if __name__ == '__main__':
     main()
