+++ ros2_ws/src/assembly_robot/yaml/well_positions.yaml
